Add CentOS support for role cephclient
