+++ roles/cephclient/defaults/main.yml
@@ -16,7 +16,7 @@ operator_group: "{{ operator_user }}"
 # cephclient
 
 cephclient_install_type: container
-cephclient_version: pacific
+cephclient_version: reef
 
 cephclient_mons: []
 cephclient_keyring: ""
@@ -43,8 +43,12 @@ cephclient_service_name: "docker-compose@cephclient"
 
 cephclient_configure_repository: true
 
-cephclient_debian_repository_key: https://download.ceph.com/keys/release.asc
+cephclient_repository_key: https://download.ceph.com/keys/release.asc
+
 cephclient_debian_repository: "deb https://download.ceph.com/debian-{{ cephclient_version }} {{ ansible_distribution_release }} main"
 
-cephclient_debian_packages:
+cephclient_redhat_arch: "{{ 'aarch64' if ansible_architecture in ['aarch64', 'arm64'] else ansible_architecture }}"
+cephclient_redhat_repository: "https://download.ceph.com/rpm-{{ cephclient_version }}/el9/{{ cephclient_redhat_arch }}/"
+
+cephclient_packages:
   - ceph
