+++ roles/cephclient/tasks/container.yml
