+++ roles/cephclient/tasks/package-Debian.yml
@@ -9,7 +9,7 @@
 - name: Add repository key with apt-key
   become: true
   ansible.builtin.apt_key:
-    url: "{{ cephclient_debian_repository_key }}"
+    url: "{{ cephclient_repository_key }}"
   when:
     - cephclient_configure_repository|bool
     - ansible_distribution_version is version('22.04', '<')
@@ -17,7 +17,7 @@
 - name: Add repository gpg key
   become: true
   ansible.builtin.get_url:
-    url: "{{ cephclient_debian_repository_key }}"
+    url: "{{ cephclient_repository_key }}"
     dest: /etc/apt/trusted.gpg.d/cephclient.asc
     mode: 0644
     owner: root
@@ -38,7 +38,7 @@
 - name: Install required packages
   become: true
   ansible.builtin.apt:
-    name: "{{ cephclient_debian_packages }}"
+    name: "{{ cephclient_packages }}"
     state: present
     lock_timeout: "{{ apt_lock_timeout | default(300) }}"
